"1.Creat FilmService  2.Creat new method in FilmController"

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -4,11 +4,11 @@ import java.util.Collection;
 import ru.yandex.practicum.filmorate.model.Film;
 
 public interface FilmStorage {
-  Film get(Integer id);
+  Film get(Long id);
 
   Film add(Film film);
 
-  Film delete(Integer id);
+  Film delete(Long id);
 
   Film update(Film film);
